chore: Update authService.js and redditServices.js
This commit updates the authService.js and redditServices.js files. It includes changes to handle token expiration and validation, as well as redirecting to authenticate again if necessary. The code also removes unused code and adds token validation checks in the fetchUserData and fetchNewPosts functions.

Issue: Will get stuck in loop of authentication.