chore: Update Header component with user login functionality
This commit adds user login functionality to the Header component. It imports the `useSelector` hook from `react-redux` and the `selectAllUserData` selector from `../../redux/features/userData`. It also adds a `userName` state variable and sets it to the first user's name from the `userData` array. The user's name is displayed in the header along with a logout button.